Refactor the structure of Eventity\Code
