--- src/Code/ClassAnalyser.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Eventity\Code;
-
-interface ClassAnalyser
-{
-    /**
-     * @param string $className
-     *
-     * @return ClassDefinition
-     */
-    public function analyse($className);
-}
