+++ src/Code/Definition/FieldDefinition.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Eventity\Code;
+namespace Eventity\Code\Definition;
 
 use Assert\Assertion;
