+++ src/Code/Instantiater/ClassInstantiater.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Eventity\Code;
+namespace Eventity\Code\Instantiater;
 
 interface ClassInstantiater
 {
