+++ src/Code/Instantiater/ReflectionClassInstantiater.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Eventity\Code\Instantiater;
+
+use Assert\Assertion;
+
+final class ReflectionClassInstantiater implements ClassInstantiater
+{
+    public function instantiate($className, array $args = [])
+    {
+        Assertion::string($className);
+
+        return new $className(...$args);
+    }
+}
