+++ src/Code/Renderer/CodeRenderer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Eventity\Code;
+namespace Eventity\Code\Renderer;
 
 interface CodeRenderer
 {
