+++ src/Code/Renderer/DefaultCodeRenderer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Eventity\Code;
+namespace Eventity\Code\Renderer;
 
 final class DefaultCodeRenderer implements CodeRenderer
 {
