+++ tests/unit/Code/Renderer/DefaultClassCodeRendererSpec.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace unit\Eventity\Code;
+namespace unit\Eventity\Code\Renderer;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Eventity\Code\ClassDefinition;
-use Eventity\Code\FieldDefinition;
-use Eventity\Code\MethodDefinition;
-use Eventity\Code\ClassCodeRenderer;
-use Eventity\Code\CodeRenderer;
+use Eventity\Code\Definition\ClassDefinition;
+use Eventity\Code\Definition\FieldDefinition;
+use Eventity\Code\Definition\MethodDefinition;
+use Eventity\Code\Renderer\ClassCodeRenderer;
+use Eventity\Code\Renderer\CodeRenderer;
 
 final class DefaultClassCodeRendererSpec extends ObjectBehavior
 {
